only give the text remaining after the route
Since we set up the route, we know what the text starts with, or we
shouldn't care. Plugins/callbacks should not need to know how they are
being routed to. If we kept the route text in the message, the callback
would have to know what to ignore.

We may add the route as a new property to the Freyr::Message object if
it becomes necessary.